Refactored code out from operations.py into utils, to be used by other folders
