+++ cli/operations.py
@@ -1,4 +1,3 @@
-import configparser
 import requests
 import time
 import os
@@ -7,20 +6,7 @@ import pathlib
 import json
 import copy
 from azure.storage.blob import BlockBlobService, ContentSettings
-
-FUNCTIONS_SECTION = 'FUNCTIONS'
-FUNCTIONS_KEY = 'FUNCTIONS_KEY'
-FUNCTIONS_URL = 'FUNCTIONS_URL'
-
-STORAGE_SECTION = 'STORAGE'
-STORAGE_KEY = 'STORAGE_KEY'
-STORAGE_ACCOUNT = 'STORAGE_ACCOUNT'
-STORAGE_CONTAINER = 'STORAGE_CONTAINER'
-
-TAGGING_SECTION = 'TAGGING'
-TAGGING_LOCATION_KEY = 'TAGGING_LOCATION'
-TAGGING_USER_KEY = 'TAGGING_USER'
-
+from blob_utils import BlobStorage
 
 DEFAULT_NUM_IMAGES = 40
 LOWER_LIMIT = 0
@@ -31,31 +17,11 @@ CONFIG_PATH = os.environ.get('ALCONFIG', None)
 azure_storage_client = None
 
 
-class MissingConfigException(Exception):
-    pass
-
-
 class ImageLimitException(Exception):
     pass
 
-
-def get_azure_storage_client(config):
-    # Todo: Move away from global client.
-    global azure_storage_client
-
-    if azure_storage_client is not None:
-        return azure_storage_client
-
-    azure_storage_client = BlockBlobService(
-        config.get("storage_account"),
-        account_key=config.get("storage_key")
-    )
-
-    return azure_storage_client
-
-
 def onboard(config, folder_name):
-    blob_storage = get_azure_storage_client(config)
+    blob_storage = BlobStorage.get_azure_storage_client(config)
     uri = 'https://' + config.get("storage_account") + '.blob.core.windows.net/' + config.get("storage_container") + '/'
     functions_url = config.get('url') + '/api/onboarding'
     images = []
@@ -238,75 +204,3 @@ def trim_file_paths(json_data):
 
     return modified_json
 
-
-def read_config(config_path):
-    if config_path is None:
-        raise MissingConfigException()
-
-    parser = configparser.ConfigParser()
-    parser.read(config_path)
-    return read_config_with_parsed_config(parser)
-
-
-def functions_config_section(functions_config_section):
-    functions_key_value = functions_config_section.get(FUNCTIONS_KEY)
-    functions_url_value = functions_config_section.get(FUNCTIONS_URL)
-
-    if not functions_key_value or not functions_url_value:
-        raise MissingConfigException()
-
-    return functions_key_value, functions_url_value
-
-
-def storage_config_section(storage_config_section):
-    storage_account_value = storage_config_section.get(STORAGE_ACCOUNT)
-    storage_key_value = storage_config_section.get(STORAGE_KEY)
-    storage_container_value = storage_config_section.get(STORAGE_CONTAINER)
-
-    if not storage_account_value or not storage_key_value or not storage_container_value:
-        raise MissingConfigException()
-
-    return storage_account_value, storage_key_value, storage_container_value
-
-
-def tagging_config_section(tagging_config_section):
-    tagging_location_value = tagging_config_section.get(TAGGING_LOCATION_KEY)
-    tagging_user_value = tagging_config_section.get(TAGGING_USER_KEY)
-
-    if not tagging_location_value or not tagging_user_value:
-        raise MissingConfigException()
-
-    return tagging_location_value, tagging_user_value
-
-
-def read_config_with_parsed_config(parser):
-    sections = parser.sections()
-
-    if FUNCTIONS_SECTION not in sections:
-        raise MissingConfigException()
-
-    if STORAGE_SECTION not in sections:
-        raise MissingConfigException()
-
-    if TAGGING_SECTION not in sections:
-        raise MissingConfigException()
-
-    functions_key, functions_url = functions_config_section(
-        parser[FUNCTIONS_SECTION]
-    )
-
-    storage_account, storage_key, storage_container = storage_config_section(
-        parser[STORAGE_SECTION]
-    )
-
-    tagging_location, tagging_user = tagging_config_section(parser[TAGGING_SECTION])
-
-    return {
-        "key": functions_key,
-        "url": functions_url,
-        "storage_account": storage_account,
-        "storage_key": storage_key,
-        "storage_container": storage_container,
-        "tagging_location": tagging_location,
-        "tagging_user": tagging_user
-    }
